requirements.txt
numpy>=1.24
torch>=2.0
matplotlib>=3.7
pyyaml>=6.0
tqdm>=4.65
